feat(input): add custom prompt input field
